signal waiting - always wait for signals before _Exit.
Centralize waiting in a single, normal exit location. This avoids
us loosing and not being able to attach to and debug a crashed
child thread when another thread decides to _Exit for some reason
while the signal handler is running in the other thread.
